test fixes
- fix: refCount never decreased on DCLTexture??
- split metrics counter tests in integration and unit classes
- add APK cleanup and rendererState = true to IntegrationTestSuite
