+++ unity-renderer/Assets/Scripts/Tests/IntegrationTestSuite/IntegrationTestSuite.asmdef
@@ -16,7 +16,8 @@
     "GUID:97d8897529779cb49bebd400c7f402a6",
     "GUID:3047fd2f05bcaae498a99a3f9321f9a8",
     "GUID:3b80b0b562b1cbc489513f09fc1b8f69",
-    "GUID:cc6bfabbfe87b7949aa248893f89ce37"
+    "GUID:cc6bfabbfe87b7949aa248893f89ce37",
+    "GUID:4720e174f2805c74bb7aa94cc8bb5bf8"
   ],
   "includePlatforms": [],
   "excludePlatforms": [],
